refactor: extract Dropdown component from Select
